写完了获得isbn，得到这种booktype下面的评论和回复的功能，但还没有测试，可能有需要修改的地方

+++ backend/src/main/java/com/example/lab2/entity/Comment.java
@@ -35,9 +35,9 @@ public class Comment {
 
     private boolean deletedByAdmin;
 
-    private int rate;
+    private long rate;
 
-    public Comment(long userID, String isbn, String content, Date time, boolean deletedBySelf, boolean deletedByAdmin, int rate) {
+    public Comment(long userID, String isbn, String content, Date time, boolean deletedBySelf, boolean deletedByAdmin, long rate) {
         this.userID = userID;
         this.isbn = isbn;
         this.content = content;
